refactor: Convert entire codebase from JSX to TSX with comprehensive TypeScript support
- Convert all 57 JSX files to TSX with proper TypeScript interfaces
- Add TypeScript interfaces for all components (AdminLayout, GlassInput, GlassButton, Heading, Button, Modal, etc.)
- Fix prop type issues: min/max/step props converted from strings to numbers
- Add proper typing for page components with Promise-based params
- Enhance Modal component to support both children and default content
- Add event handler type casting for proper TypeScript support
- Remove legacy error.js file in favor of comprehensive error.ts
- Ensure full build compatibility with zero TypeScript errors

This completes Phase 1 refactoring with comprehensive TypeScript integration
while maintaining all existing functionality and backward compatibility.
