+++ src/app/components/Footer/FooterLogo.tsx
