+++ src/app/components/Glassmorphism/GlassInput.tsx
@@ -1,6 +1,21 @@
 "use client";
 import React from 'react';
 
+interface GlassInputProps {
+  type?: string;
+  value: any;
+  onChange: any;
+  placeholder?: string;
+  className?: string;
+  disabled?: boolean;
+  required?: boolean;
+  maxLength?: number;
+  min?: number;
+  max?: number;
+  step?: number;
+  [key: string]: any;
+}
+
 const GlassInput = ({
   type = "text",
   value,
@@ -14,7 +29,7 @@ const GlassInput = ({
   max,
   step,
   ...props
-}) => {
+}: GlassInputProps) => {
   const baseClasses = "w-full px-4 py-3 bg-white/10 border border-white/20 rounded-xl text-white placeholder-white/80 focus:outline-none focus:border-[#3a8fff] transition-colors";
   
   return (
