+++ src/app/components/HeroSection/HeroBackground.tsx
