+++ src/app/components/HeroSection/HeroContent.tsx
