+++ src/app/components/HeroSection/HeroLogo.tsx
