+++ src/app/components/HiTechSection/HiTechSection.tsx
