+++ src/app/components/NeonBubblesBackground.tsx
