+++ src/app/components/common/StockStatusBadge.tsx
