+++ src/app/components/ui/ProductCard.tsx
@@ -115,6 +115,8 @@ const ProductCard = ({
                   color="#0A6EBD"
                   bgColor="#fff"
                   size="medium"
+                  padding="px-4 py-2"
+                  textSize="text-sm"
                   className="w-full max-w-[140px] mx-auto mt-2"
                   href={isOutOfStock ? undefined : (buttonHref || `/products/${product.categorySlug || product.category.toLowerCase()}/${encodeURIComponent(product.catNo)}`)}
                   onClick={buttonOnClick}
