+++ src/app/products/[category]/[product]/VariantCarousel.tsx
